Add CLI to mock_authority
Since a convenient way to get a blind signature for a given blinded digital signature PK is required for testing the client, add that functionality to the mock authority binary as a CLI command. Also separate process related code like STDIO reading to separate subcrate.
